DRY config parsing, improve error handling, ERB -> Erubis!
Erubis:  You get it for free when you install chef-omnibus and it's
simpler to render a template with it.  We now build templates with
Kitchenplan's ohai data object in the "@node" namespace... so you
can write your Kitchenplan configs like they're Chef templates!
Note that we aren't loading all the Ohai plugins at the moment.