A little more debug printing.  (YAML parser still yields nil?!)
This was all tested as working locally on Mavericks, so I'm not sure
why it's decided to stop working in Travis.  Perhaps it's an
RVM vs. System Ruby issue?  Could Erubis be the culprit?  Is it a
permissions issue on the host?  Let's break the parsing down into
steps and print the results of each one...

(This should probably be reverted once the problem step is identified)